Hide Expand All/None if no expandable collections
